node: streamline scripts for building node
It seems like just `prepare` should be enough for everything.
