--- src/clients/node/scripts/build.js
@@ -1,4 +1,0 @@
-const { execSync } = require('child_process')
-
-execSync('npm run build_tsc')
-execSync('npm run build_lib')
